use fixture in tests_common module
